requirements.txt
pandas==2.2.3
numpy==2.3.5
scikit-learn==1.8.0
matplotlib==3.10.8
joblib==1.5.3
PyYAML==6.0.3
fastapi==0.111.0
uvicorn[standard]==0.40.0
pydantic==2.11.7
requests==2.32.4
prometheus-client==0.22.1
mlflow>=3.0,<4.0
dvc>=3.0,<4.0
evidently>=0.7,<0.8
jupyterlab>=4.0,<5.0
pytest==8.4.1
httpx==0.28.1
